# packages/sdk/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Lint API
  version: 1.0.0
  description: |
    Public API surface shared by:
      - Lint Cloud (lint.to / api.lint.to) — managed SaaS, hosts policies and team data
      - @lint/server — local Hono server embedded in @lint/desktop and runnable standalone

    @osmove/lint-sdk is the typed TypeScript client for both. End-users point
    `createLintClient({ baseUrl })` at whichever instance they're talking to.

    Regenerate types after editing this file:
      pnpm --filter @osmove/lint-sdk openapi:fetch    # (cloud only)
      pnpm --filter @osmove/lint-sdk openapi:types
servers:
  - url: https://api.lint.to
    description: Lint Cloud (production)
  - url: http://127.0.0.1:7878
    description: Local @lint/server
...
paths:
...
  /api/repos:
    get:
      tags: [repos]
      summary: List registered repositories (registry merged with workspace)
      operationId: listRepos
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items: { $ref: "#/components/schemas/Repo" }
    post:
      tags: [repos]
      summary: Register a repository in ~/.lint/projects.json
      operationId: createRepo
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string }
                name: { type: string }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RepoEntry" }
  /api/projects:
    get:
      tags: [projects]
      summary: List projects with repository children
      operationId: listProjects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
    post:
      tags: [projects]
      summary: Create a project, optionally seeded with a repository root
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                root: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /api/projects/{id}:
    get:
      tags: [projects]
      summary: Get a project by id
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { description: Not found }
  /api/projects/{id}/repos:
    post:
      tags: [projects]
      summary: Attach a repository to a project
      operationId: addProjectRepo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string }
                name: { type: string }
      responses:
        "201":
          description: Attached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repo" }
        "404": { description: Not found }
  /api/projects/{id}/repos/{repoId}:
    delete:
      tags: [projects]
      summary: Remove a repository from a project
      operationId: removeProjectRepo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: repoId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404": { description: Not found }
  /api/repos/{id}:
    delete:
      tags: [repos]
      summary: Unregister a repository
      operationId: removeRepo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404": { description: Not found }
...
components:
  schemas:
...
    Run:
      type: object
      required: [id, startedAt, errorCount, warningCount, status]
      properties:
        id:           { type: string }
        startedAt:    { type: string, format: date-time }
        finishedAt:   { type: string, format: date-time, nullable: true }
        errorCount:   { type: integer, minimum: 0 }
        warningCount: { type: integer, minimum: 0 }
        status:
          type: string
          enum: [running, passed, failed]
...
    RepoEntry:
      type: object
      required: [id, name, root, addedAt]
      properties:
        id:      { type: string }
        name:    { type: string }
        root:    { type: string }
        addedAt: { type: string, format: date-time }
    Project:
      type: object
      required: [id, name, ephemeral, repositories]
      properties:
        id:        { type: string }
        name:      { type: string }
        root:      { type: string, nullable: true }
        addedAt:   { type: string, nullable: true, format: date-time }
        ephemeral: { type: boolean }
        repositories:
          type: array
          items: { $ref: "#/components/schemas/Repo" }
    Repo:
      type: object
      required: [id, name, root, projectId, projectName, ephemeral, health]
      properties:
        id:        { type: string }
        name:      { type: string }
        root:      { type: string }
        projectId: { type: string }
        projectName: { type: string }
        addedAt:   { type: string, nullable: true, format: date-time }
        ephemeral: { type: boolean }
        health:
          type: string
          enum: [unknown, passed, failed]
        latestRun:
          oneOf:
            - { $ref: "#/components/schemas/Run" }
            - { type: "null" }
        summary:
          type: object
          properties:
            total:   { type: integer, minimum: 0 }
            passed:  { type: integer, minimum: 0 }
            failed:  { type: integer, minimum: 0 }
            running: { type: integer, minimum: 0 }
...